+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -167,7 +167,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-11-13T09:21:05.843Z",
+    "2025-11-13T11:08:17.262Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -207,7 +207,7 @@ export const CreateContactResponseBody$outboundSchema: z.ZodType<
   tags: z.array(z.string()).optional(),
   metadata: z.record(z.any()).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-11-13T09:21:05.843Z"))
+  updated: z.date().default(() => new Date("2025-11-13T11:08:17.262Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {